+++ test/correctness/PySys_internal_183/Input/MyNestedTestcase/pysystest.py
@@ -1,0 +1,25 @@
+__pysys_title__   = r""" Category TODO - Test title (short, specific and sortable!) goes here TODO """ 
+#                        ================================================================================
+__pysys_purpose__ = r""" """ 
+	
+__pysys_authors__ = "bsp"
+__pysys_created__ = "2022-03-10"
+#__pysys_skipped_reason__   = "Skipped until Bug-1234 is fixed"
+
+#__pysys_traceability_ids__ = "Bug-1234, UserStory-456" 
+#__pysys_groups__           = "myGroup, disableCoverage, performance"
+#__pysys_modes__            = lambda helper: helper.inheritedModes + [ {'mode':'MyMode', 'myModeParam':123}, ]
+#__pysys_parameterized_test_modes__ = {'MyParameterizedSubtestModeA':{'myModeParam':123}, 'MyParameterizedSubtestModeB':{'myModeParam':456}, }
+
+import pysys
+from pysys.constants import *
+
+import os, sys, math, shutil, glob
+
+class PySysTest(pysys.basetest.BaseTest):
+
+	def execute(self):
+		pass
+		
+	def validate(self):
+		pass